requirements.txt
python-telegram-bot==21.10
requests==2.31.0
python-dotenv==1.0.0
aiosqlite==0.20.0
cryptography==44.0.0
httpx==0.27.0